test(httpruntime): split HTTPRuntimeTests to mirror its source files
