--- Tests/HTTPRuntimeTests/HTTPRuntimeTests.swift
@@ -1,111 +1,0 @@
-//
-//  HTTPRuntimeTests.swift
-//  HTTPRuntime
-//
-//  Created by Guilherme Souza on 08/07/26.
-//
-
-import Foundation
-import HTTPTypes
-import Testing
-
-@testable import HTTPRuntime
-
-@Suite
-struct HTTPRuntimeTests {
-
-  @Test
-  func multipartAssemblesToFileWithoutBufferingSource() throws {
-    let sourceURL = FileManager.default.temporaryDirectory
-      .appendingPathComponent("src-\(UUID().uuidString).bin")
-    let payload = Data((0..<200_000).map { UInt8($0 % 256) })
-    try payload.write(to: sourceURL)
-    defer { try? FileManager.default.removeItem(at: sourceURL) }
-
-    let form = MultipartFormData(boundary: "TESTBOUNDARY")
-      .addText(name: "meta", value: #"{"k":"v"}"#)
-      .addFile(
-        name: "file", fileURL: sourceURL, fileName: "big.bin", mimeType: "application/octet-stream")
-
-    let bodyURL = try form.buildToTempFile()
-    defer { try? FileManager.default.removeItem(at: bodyURL) }
-    let body = try Data(contentsOf: bodyURL)
-
-    #expect(form.contentType == "multipart/form-data; boundary=TESTBOUNDARY")
-    let text = String(decoding: body.prefix(400), as: UTF8.self)
-    #expect(text.contains("--TESTBOUNDARY"))
-    #expect(text.contains(#"Content-Disposition: form-data; name="meta""#))
-    #expect(text.contains(#"name="file"; filename="big.bin""#))
-    #expect(body.count > payload.count)
-  }
-
-  @Test
-  func addHeaderAppendsToExistingValue() throws {
-    var builder = HTTPRequestBuilder(
-      method: .get, baseURL: URL(string: "https://example.com")!, path: "/x")
-    builder.addHeader(.prefer, value: "returning=minimal")
-    builder.addHeader(.prefer, value: "count=exact")
-    let request = try builder.build()
-    #expect(request.headerFields[.prefer] == "returning=minimal,count=exact")
-  }
-
-  @Test
-  func addHeaderReplacesMatchingDirectiveKey() throws {
-    var builder = HTTPRequestBuilder(
-      method: .get, baseURL: URL(string: "https://example.com")!, path: "/x")
-    builder.addHeader(.prefer, value: "count=exact")
-    builder.addHeader(.prefer, value: "returning=minimal")
-    builder.addHeader(.prefer, value: "count=planned")
-    let request = try builder.build()
-    #expect(request.headerFields[.prefer] == "count=planned,returning=minimal")
-  }
-
-  @Test
-  func addHeaderSetsWhenAbsent() throws {
-    var builder = HTTPRequestBuilder(
-      method: .get, baseURL: URL(string: "https://example.com")!, path: "/x")
-    builder.addHeader(.prefer, value: "returning=minimal")
-    let request = try builder.build()
-    #expect(request.headerFields[.prefer] == "returning=minimal")
-  }
-
-  @Test
-  func addHeaderIgnoresNilValue() throws {
-    var builder = HTTPRequestBuilder(
-      method: .get, baseURL: URL(string: "https://example.com")!, path: "/x")
-    builder.addHeader(.prefer, value: "returning=minimal")
-    builder.addHeader(.prefer, value: nil)
-    let request = try builder.build()
-    #expect(request.headerFields[.prefer] == "returning=minimal")
-  }
-
-  @Test
-  func pathEncoding() {
-    #expect(PathEncoding.segment("a/b c") == "a%2Fb%20c")
-    #expect(PathEncoding.greedy("a/b/c.txt") == "a/b/c.txt")
-    #expect(PathEncoding.greedy("a/b c.txt") == "a/b%20c.txt")
-  }
-
-  @Test
-  func jsonValueRoundTrip() throws {
-    let value = JSONValue.object([
-      "s": .string("x"),
-      "n": .number(3.5),
-      "b": .bool(true),
-      "arr": .array([.number(1), .null]),
-    ])
-    let data = try JSONCoding.encoder.encode(value)
-    let decoded = try JSONCoding.decoder.decode(JSONValue.self, from: data)
-    #expect(decoded == value)
-  }
-
-  @Test
-  func iso8601DateCoding() throws {
-    struct Holder: Codable, Equatable { let at: Date }
-    let json = #"{"at":"2026-07-06T12:34:56.789Z"}"#
-    let decoded = try JSONCoding.decoder.decode(Holder.self, from: Data(json.utf8))
-    let reencoded = try JSONCoding.encoder.encode(decoded)
-    let round = try JSONCoding.decoder.decode(Holder.self, from: reencoded)
-    #expect(decoded == round)
-  }
-}
